backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Face Recognition (IMPORTANT: Install these in specific order)
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
tensorflow==2.15.0
tf-keras==2.15.0
deepface==0.0.79

# Vector Search
faiss-cpu==1.7.4

# Database
motor==3.3.2
pymongo==4.6.0

# Utils
python-dotenv==1.0.0
aiofiles==23.2.1